unpack scripts update.

+++ android/UnpackUnityAndroid.sh
@@ -4,6 +4,17 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 unzipFolder="${SCRIPT_DIR}/unityLibrary"
 
+#Check for checksum
+checkSumFile="${SCRIPT_DIR}/unity_zip_check_sum"
+checksum=$(shasum -a256 "${SCRIPT_DIR}/unityLibrary.zip")
+oldChecksum=`cat "$checkSumFile"`
+
+#If the same exit
+if [ "$checksum" = "$oldChecksum" ]; then
+    exit
+fi
+
+#If not unzip it
 if [ -d "$unzipFolder" ]; then
   rm -r "$unzipFolder"
 fi
@@ -14,11 +25,6 @@ unzip -o "${SCRIPT_DIR}/unityLibrary.zip" -d "$unzipFolder"
 unityGradlePath="${unzipFolder}/build.gradle"
 value=`cat "$unityGradlePath"`
 updated_gradle=$value
-#updated_gradle="${updated_gradle}
-#    buildTypes {
-#        profile {}
-#    }
-#}"
 
 if [[ "$OSTYPE" == "darwin"* ]]; then
     updated_gradle=${updated_gradle//\/il2cpp.exe/"/il2cpp"}
@@ -32,6 +38,7 @@ fi
 echo "$updated_gradle" > "$unityGradlePath"
 echo "$updated_gradle"
 
+#Update il2cpp
 il2cpp_path="${unzipFolder}/src/main/Il2CppOutputProject/IL2CPP/build/deploy/netcoreapp3.1"
 rm -r "$il2cpp_path"
