+++ ios/UnpackUnityIos.sh
@@ -4,6 +4,17 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 unzipFolder="${SCRIPT_DIR}/UnityLibrary"
 
+#Check for checksum
+checkSumFile="${SCRIPT_DIR}/unity_zip_check_sum"
+checksum=$(shasum -a256 "${SCRIPT_DIR}/UnityLibrary.zip")
+oldChecksum=`cat "$checkSumFile"`
+
+#If the same exit
+if [ "$checksum" = "$oldChecksum" ]; then
+    exit
+fi
+
+#If not unzip it
 if [ -d "$unzipFolder" ]; then
   rm -r "$unzipFolder"
 fi
